refactor: rename files to follow kebab-case naming convention
