+++ src/modules/users/dto/response.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 
 import { IsEmail, IsNotEmpty, IsString, IsBoolean } from 'class-validator';
 
-export class ResponseUserDTO {
+export class ResponseUserDto {
   @ApiProperty({
     example: '61c674384-f944-401b-949b-b76e8793bdc9',
     description: 'The ID of the user',
